Added wrapper functions and a loop that combines derivs
Created the high level functions that the user will use, and a class to contain the gradient results.

Added the loop into the backprop that combines the forwardprop and backprpop tensors. This deviates from my private version of the code where this loop was entirely seperate to the forwardprop and backprop. The reason for this change is to allow me to delete the forwardprop tensors as I generate the backprop. This reduceds the amount of memory needed for the adjoint method by 50%.