Degeneracy setup for TEMPO
Boolean attribute unique added to the tempo class, that if true will use degeneracy checking.

This is done by a new function called row_degeneracy (Taken from original TEMPO publication). These are then stored as attributes for the different directions.

The backend is then changed such that the legs of the rank 4 influence tensors have the right size and values.