Redid algorithm, and introduced extra node to recombination
Basically made the previous algorithm i introduced defunct as i'm not convinced it
adds any time saving that would warrent the increase in complexity and modes of
failure of the code.

Instead added the extra node necessary during the chain rule, which i previously
forgot to add.

Whole thing is still bug, one is because i haven't treated the first and last
half propagator seperately as they're a special case, but there is more bugs
that i don't know about