Optional system argument for guess_tempo_parameters
If system is provided, the maximum system frequency (from spectral norm of
Hamiltonian or Dissipators) is sued to estimate a suitable timestep size
dt to calculate the system-only dynamics. If this is smaller than that
estimated from the bath correlation functions, then it is returned
instead.

Also centred plot axes of plot_correlations_with_parameters to make it
easier to identify decay to 0